exported adapting to client function to utils file
